Refactor game related code
